+++ CHANGES.rst
@@ -15,6 +15,9 @@ CHANGES
 - Remove deprecatared __of__ for BrowserViews
   [MrTango]
 
+- Switch test setup to use plone.testing
+
+
 1.1 (2012-08-30)
 ----------------